feat(): removeItem error handle and empty list ui

+++ back/services/userService.js
@@ -1,7 +1,17 @@
 import usersRepository from "../repositories/usersRepository.js";
 
 const getUserById = async (id) => {
-  return await usersRepository.getUser(id);
+  const user = await usersRepository.getUser(id);
+
+  if (!user) {
+    return await usersRepository.createUser(id);
+  }
+
+  return user;
+};
+
+const deleteUser = async (userId) => {
+  return await usersRepository.deleteUser(userId);
 };
 
 const addProduct = async (userId, product) => {
@@ -13,6 +23,7 @@ const editProduct = async (userId, product) => {
 };
 
 const removeProduct = async (userId, productId) => {
+  if (!(await isProductExists(userId, productId))) throw Error("no product");
   return await usersRepository.removeProductFromUsersList(userId, productId);
 };
 
@@ -20,8 +31,15 @@ const deleteProduct = async (productId) => {
   return await usersRepository.deleteProductFromDB(productId);
 };
 
+const isProductExists = async (userId, productId) => {
+  return (await usersRepository.getUser(userId)).products.hasOwnProperty(
+    productId
+  );
+};
+
 export default {
   getUserById,
+  deleteUser,
   addProduct,
   editProduct,
   removeProduct,
